SC1117: Backslash is literal in "\X". Prefer explicit escaping: "\\X".
In a double quoted string, you have escaped a character that has no
special behavior when escaped. Instead, it's invoking the fallback
behavior of being interpreted literally.

Instead of relying on this implicit fallback, you should escape the
backslash explicitly. This makes it clear that it's meant to be passed
as a literal backslash in the string parameter.