SC2064: Use single quotes, otherwise this expands now rather than when signalled.
With double quotes, all parameter and command expansions will expand
when the trap is defined rather than when it's executed.

In the example, the message will contain the date on which the trap was
declared, and not the date on which the script exits.

Using single quotes will prevent expansion at declaration time, and save
it for execution time.