Add automatic metapkg support for Python flavors.
When a Python port has flavors, it generates many packages, py27-foo,
py35-foo, py36-foo... and sometime, you don't care which you are
installing you only want the current one.

With this, a py-foo meta package gets created, and it will always depend
on the current default Python version of the package.