svnadmin/conf/mentors: Graduate Kyle!
Under mentorship, Kyle has successfuly demonstrated a high-quality
approach to contribution to the FreeBSD Ports ecosystem, including, but not
limited to:

  - Sweeping the ports tree for future regex compliance issues
  - VuXML entry for new CVEs
  - Many merges back to quarterly branches
  - More delicate surgery to bring in the new version of qemu-user-static

Otherwise more than satisfying:

   https://wiki.freebsd.org/KubilayKocak/MentoringChecklist

This commit [1] graduates Kyle, congratulations!

[1] This had originally been designed as a self-graduation process, with
Kyle putting the final touch on a process he was intentionally asked to
drive (creator of ones own destiny). This caused a deadlock with the
tradition of mentors releasing their mentees.

In the interests of making forward progress in this instance, I will commit
this change, but I/we will cross this bridge again with future mentees.

Submitted by:		kevans (mentee)
Requested by:		kevans (mentee)
Differential Revision:	D25395